I have combined 18 commits concern with db_init. When I was coding db_init task, I have broken this task into mangable mini-task. However, this db_init task should be just one change because there will be error if you do multiple mini-changes.
This is a combination of 17 commits.

The first commit's message is:

1) database connection is added for mysql dev
2) test for db init
3) db init in migration and create "news" table migration script is
created.
4) unit tested OK

This is the 2nd commit message:

- unit test of create album table migration script
- create album table migration script

This is the 3rd commit message:

- unit test for create photo table migration
- create photo table in migration script

This is the 4th commit message:

As Laravel migrate return column rather than table, my chaining of add
column does not work and it created only the first column.

$table->increments('id')
        ->text('description')
        ->integer('creator');

I have chanined the column adding like above and only "id" column got
created. This is because the column creation method return the column
object in Laravel rather than table. I have change my migration script
to stop chaining (see below) and all columns are created when you run
migration script.

$table->increments('id');
$table->text('description');
$table->integer('creator');

This is the 5th commit message:

- unit test for create video table migration script
- create video table in migration script

This is the 6th commit message:

- start putting namespace Tisarana\Test in test and corrected errors of
  using namespace
- check the columns existance in table creation test not just table
  existance

This is the 7th commit message:

- unit test for create comment table migration script
- create comment table migration script

This is the 8th commit message:

- unit test for create event table migration script
- create event table migration script

This is the 9th commit message:

- unit test for create user table migration script
- create user table migration script

This is the 10th commit message:

- unit test for create group table migration script
- create group table migration script

This is the 11th commit message:

- unit test for create group_memebr table migration script
- create group_member table migration script

This is the 12th commit message:

- group table name field has cut and copy error of length 1500 and
  corrected back to 50

This is the 13th commit message:

- unit test for create object table in migration script
- create object table in migration script

This is the 14th commit message:

- unit test for create action table in migration script
- create action table in migration script

This is the 15th commit message:

- unit test for create acl table in migration script
- create acl table in migration script

This is the 16th commit message:

- remove "default(0)" in "creator" fields of all tables except comment
- remove "type" field from comment
- add "link_object" field in comment
- correct length of "name" field in event to 50
- add indexes to all tables

This is the 17th commit message:

- add compound primary key (principal, principal_type, action columns)
  in acl table
- remove primary key id and add compund primary key (group, user
  columns) in group_member table